+++ src/lib/mympd_state.c
@@ -16,6 +16,9 @@
 #include "src/lib/utility.h"
 #include "src/mpd_client/presets.h"
 #include "src/mympd_api/home.h"
+#ifdef MYMPD_ENABLE_LUA
+    #include "src/mympd_api/scripts/vars.h"
+#endif
 #include "src/mympd_api/timer.h"
 #include "src/mympd_api/trigger.h"
 
@@ -27,6 +30,14 @@
  * @param free_data true=free the struct, else not
  */
 void mympd_state_save(struct t_mympd_state *mympd_state, bool free_data) {
+    // write album cache to disc
+    // only for simple mode to save the cached uris
+    if (mympd_state->config->save_caches == true &&
+        mympd_state->config->albums.mode == ALBUM_MODE_SIMPLE)
+    {
+        album_cache_write(&mympd_state->album_cache, mympd_state->config->workdir,
+            &mympd_state->mpd_state->tags_album, &mympd_state->config->albums, true);
+    }
     struct t_partition_state *partition_state = mympd_state->partition_state;
     while (partition_state != NULL) {
         last_played_file_save(partition_state);
@@ -35,6 +46,9 @@ void mympd_state_save(struct t_mympd_state *mympd_state, bool free_data) {
     }
     mympd_api_home_file_save(&mympd_state->home_list, mympd_state->config->workdir);
     mympd_api_timer_file_save(&mympd_state->timer_list, mympd_state->config->workdir);
+    #ifdef MYMPD_ENABLE_LUA
+        mympd_api_script_vars_file_save(&mympd_state->script_var_list, mympd_state->config->workdir);
+    #endif
     mympd_api_trigger_file_save(&mympd_state->trigger_list, mympd_state->config->workdir);
     if (free_data == true) {
         mympd_state_free(mympd_state);
@@ -113,6 +127,8 @@ void mympd_state_default(struct t_mympd_state *mympd_state, struct t_config *con
     list_init(&mympd_state->home_list);
     //timer
     mympd_api_timer_timerlist_init(&mympd_state->timer_list);
+    //variables for scripts
+    list_init(&mympd_state->script_var_list);
     //album cache
     cache_init(&mympd_state->album_cache);
     //init last played songs list
@@ -132,6 +148,8 @@ void mympd_state_free(struct t_mympd_state *mympd_state) {
     list_clear(&mympd_state->home_list);
     //timer
     mympd_api_timer_timerlist_clear(&mympd_state->timer_list);
+    //variables for scripts
+    list_clear(&mympd_state->script_var_list);
     //mpd shared state
     mpd_state_free(mympd_state->mpd_state);
     //partition state