+++ src/lib/mympd_state.h
@@ -223,6 +223,7 @@ struct t_mympd_state {
     struct t_timer_list timer_list;               //!< list of timers
     struct t_list home_list;                      //!< list of home icons
     struct t_list trigger_list;                   //!< list of triggers
+    struct t_list script_var_list;                //!< list of variables for scripts
     sds tag_list_search;                          //!< comma separated string of tags for search
     sds tag_list_browse;                          //!< comma separated string of tags for browse
     bool smartpls;                                //!< enable smart playlists